+++ CHANGELOG.md
@@ -56,6 +56,8 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
 - Expose `argmax()` directly in _util_ module
 
 ## 🏥 Fixes
+- Fixed `n_buffered()` method of the Postgres statistics interface raising an error if no pages of the relation are currently
+  buffered. We now return 0 in this case.
 - Fixed string representation of `COUNT(DISTINCT ...)` for multiple arguments. We now generate the correct
   `COUNT(DISTINCT (a, b))` instead of `COUNT(DISTINCT a, b)`.
 - Fixed output format of the benchmarking log if additional entries are appended to an existing log. Essentially, we fix